move error and retry to ExportProcessorOps
