+++ core/src/modules/ops/exportProcessor.ts
@@ -1,8 +1,11 @@
+import Moment from "moment";
 import { Export } from "../../models/Export";
 import { Destination } from "../../models/Destination";
 import { CLS } from "../cls";
 import { Op } from "sequelize";
 import { ExportProcessor } from "../../models/ExportProcessor";
+import { config } from "actionhero";
+import { plugin } from "../plugin";
 
 export namespace ExportProcessorOps {
   const defaultExportProcessingDelay = 1000 * 60 * 5;
@@ -63,5 +66,80 @@ export namespace ExportProcessorOps {
     }
 
     return exportProcessors.length;
+  }
+
+  export async function setError(
+    instance: ExportProcessor,
+    error: Error,
+    retryDelay: number = config.tasks.timeout
+  ) {
+    const maxAttempts = parseInt(
+      (await plugin.readSetting("core", "export-processors-max-retries-count"))
+        .value
+    );
+
+    instance.errorMessage = error.message || error.toString();
+    if (error["errorLevel"]) instance.errorLevel = error["errorLevel"];
+
+    instance.retryCount++;
+    if (instance.retryCount >= maxAttempts) {
+      instance.state = "failed";
+      instance.processAt = null;
+      await Export.update(
+        {
+          state: "failed",
+          errorMessage: `An error occurred while processing the export: ${instance.errorMessage}`,
+          errorLevel: "error",
+        },
+        { where: { state: "processing", exportProcessorId: instance.id } }
+      );
+    } else if (instance.errorLevel === "info") {
+      instance.state = "failed";
+      await Export.update(
+        {
+          state: "failed",
+          errorMessage: instance.errorMessage,
+          errorLevel: "info",
+        },
+        { where: { state: "processing", exportProcessorId: instance.id } }
+      );
+    } else {
+      instance.processAt = Moment().add(retryDelay, "ms").toDate();
+      instance.startedAt = null;
+    }
+
+    return instance.save();
+  }
+
+  export async function retry(
+    instance: ExportProcessor,
+    retryDelay: number = config.tasks.timeout
+  ) {
+    const maxAttempts = parseInt(
+      (await plugin.readSetting("core", "export-processors-max-retries-count"))
+        .value
+    );
+
+    instance.retryCount++;
+    if (instance.retryCount >= maxAttempts) {
+      instance.state = "failed";
+      instance.processAt = null;
+
+      // fail the related exports too
+      await Export.update(
+        {
+          state: "failed",
+          errorMessage:
+            "The maximum amount of retries was reached when trying to process this export.",
+          errorLevel: "info",
+        },
+        { where: { state: "processing", exportProcessorId: instance.id } }
+      );
+    } else {
+      instance.processAt = Moment().add(retryDelay, "ms").toDate();
+      instance.startedAt = null;
+    }
+
+    return instance.save();
   }
 }
